Task9

+++ search_algorithms/0-linear.c
@@ -11,11 +11,11 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; array && i < size; i++)
 	{
-		printf("Value checked array[%d] = [%d]\n", i, array[i]);
+		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 			return (i);
 	}
